// src/app/models/Paciente.ts
export interface Paciente {
    dni_paciente?: string;
    nombre?: string;
    apellido?: string;
    sexo?: string;
    religion?: string;
    seguro?: string;
    domicilio?: string;
    fecha_nac?: string;
    edad?: number;
    edad_gesta?: string;
    dni_usuario?: string;
    idzona?: number;
    resultado_ini?: string;
    resultado_fin?: string
}